Finalise structure and clean things up (#38)
* Use new querists in checker; run it on dev pushes

* Add coverage to CI & fix typo in constants wflow

* Remove interface from __init__

* Fill out the rest of the QA checklist
